Move more components into dashboard
