+++ frontend/src/metabase/dashboard/components/DashboardSubscriptionsSidebar/AddEditSidebar/AddEditSidebar.tsx
@@ -1,9 +1,7 @@
-import {
-  getParameters,
-  getTabHiddenParameterSlugs,
-} from "metabase/dashboard/selectors";
 import type { State } from "metabase/redux/store";
 import { connect } from "metabase/utils/redux";
+
+import { getParameters, getTabHiddenParameterSlugs } from "../../../selectors";
 
 import { AddEditEmailSidebar as AddEditEmailSidebarComponent } from "./AddEditEmailSidebar";
 import { AddEditSlackSidebar as AddEditSlackSidebarComponent } from "./AddEditSlackSidebar";
