+++ frontend/src/metabase/dashboard/components/DashboardSubscriptionsSidebar/AddEditSidebar/CaveatMessage/CaveatMessage.tsx
