+++ frontend/src/metabase/dashboard/components/DashboardSubscriptionsSidebar/NewPulseSidebar.tsx
