+++ frontend/src/metabase/dashboard/components/DashboardSubscriptionsSidebar/PulsesListSidebar.styled.tsx
